+++ src/components/login/mobile/MLogin.tsx
@@ -2,9 +2,13 @@ import React from 'react';
 import LoginImage from '../../../img/login/Login.png';
 import * as ML from './MLogin.style';
 import MMoreInfo from './MMoreInfo';
-import MLoginComplete from './MLoginComplete';
 
-const MLogin = () => {
+interface LoginProps {
+    google_login: () => void,
+    kakao_login: () => void
+  }
+
+const MLogin = ({google_login, kakao_login}:LoginProps) => {
     return (
         <>
             <ML.Container>
@@ -16,16 +20,14 @@ const MLogin = () => {
                 <ML.Item src={LoginImage} alt="login" />
 
                 <ML.Box>
-                    <ML.GoogleBtn>
+                    <ML.GoogleBtn onClick={google_login}>
                         <span>Google 계정으로 로그인</span>
                     </ML.GoogleBtn>
-                    <ML.KakaoBtn>
+                    <ML.KakaoBtn onClick={kakao_login}>
                         <span>카카오 로그인</span>
                     </ML.KakaoBtn>
                 </ML.Box>
             </ML.Container>
-            <MMoreInfo />
-            <MLoginComplete />
         </>
     );
 };
